Moves drops to data.drops in preparation for moving io.py.
