+++ daliuge-engine/test/lifecycle/test_dlm.py
@@ -33,8 +33,8 @@ import unittest
 
 from dlg.ddap_protocol import DROPStates, DROPPhases
 from dlg.drop import BarrierAppDROP
-from dlg.data.directorycontainer import DirectoryContainer
-from dlg.data.file import FileDROP
+from dlg.data.drops.directorycontainer import DirectoryContainer
+from dlg.data.drops.file import FileDROP
 from dlg.droputils import DROPWaiterCtx
 from dlg.lifecycle import dlm
