+++ daliuge-engine/test/test_environmentvars.py
@@ -24,7 +24,7 @@ import tempfile
 import unittest
 
 from dlg.drop import AbstractDROP
-from dlg.data.file import FileDROP
+from dlg.data.drops.file import FileDROP
 from dlg.environmentvar_drop import EnvironmentVarDROP
 from dlg.utils import getDlgDir
