+++ daliuge-engine/test/test_graph_loader.py
@@ -24,14 +24,13 @@ import json
 import pkg_resources
 
 from dlg import graph_loader
-from dlg.common import Categories
 from dlg.ddap_protocol import DROPLinkType, DROPRel
 from dlg.drop import (
     ContainerDROP,
     AppDROP,
 )
-from dlg.data.memory import InMemoryDROP, SharedMemoryDROP
-from dlg.data.directorycontainer import DirectoryContainer
+from dlg.data.drops.memory import InMemoryDROP, SharedMemoryDROP
+from dlg.data.drops.directorycontainer import DirectoryContainer
 from dlg.common import Categories
 from dlg.apps.simple import RandomArrayApp
 
@@ -77,7 +76,7 @@ class TestGraphLoader(unittest.TestCase):
             {
                 "oid": "B",
                 "type": "container",
-                "container": "dlg.data.DirectoryContainer",
+                "container": "dlg.data.drops.DirectoryContainer",
                 "children": ["A"],
                 "dirname": ".",
             },
